Cn 6.5

+++ MareSynchronos/WebAPI/Files/FileDownloadManager.cs
@@ -16,9 +16,10 @@ namespace MareSynchronos.WebAPI.Files;
 public partial class FileDownloadManager : DisposableMediatorSubscriberBase
 {
     private readonly Dictionary<string, FileDownloadStatus> _downloadStatus;
-    private readonly FileCacheManager _fileDbManager;
     private readonly FileCompactor _fileCompactor;
+    private readonly FileCacheManager _fileDbManager;
     private readonly FileTransferOrchestrator _orchestrator;
+    private readonly List<ThrottledStream> _activeDownloadStreams;
 
     public FileDownloadManager(ILogger<FileDownloadManager> logger, MareMediator mediator,
         FileTransferOrchestrator orchestrator,
@@ -28,13 +29,33 @@ public partial class FileDownloadManager : DisposableMediatorSubscriberBase
         _orchestrator = orchestrator;
         _fileDbManager = fileCacheManager;
         _fileCompactor = fileCompactor;
+        _activeDownloadStreams = [];
+
+        Mediator.Subscribe<DownloadLimitChangedMessage>(this, (msg) =>
+        {
+            if (!_activeDownloadStreams.Any()) return;
+            var newLimit = _orchestrator.DownloadLimitPerSlot();
+            Logger.LogTrace("Setting new Download Speed Limit to {newLimit}", newLimit);
+            foreach (var stream in _activeDownloadStreams)
+            {
+                stream.BandwidthLimit = newLimit;
+            }
+        });
     }
 
-    public List<DownloadFileTransfer> CurrentDownloads { get; private set; } = new();
+    public List<DownloadFileTransfer> CurrentDownloads { get; private set; } = [];
 
     public List<FileTransfer> ForbiddenTransfers => _orchestrator.ForbiddenTransfers;
 
     public bool IsDownloading => !CurrentDownloads.Any();
+
+    public static void MungeBuffer(Span<byte> buffer)
+    {
+        for (int i = 0; i < buffer.Length; ++i)
+        {
+            buffer[i] ^= 42;
+        }
+    }
 
     public void CancelDownload()
     {
@@ -44,7 +65,7 @@ public partial class FileDownloadManager : DisposableMediatorSubscriberBase
 
     public async Task DownloadFiles(GameObjectHandler gameObject, List<FileReplacementData> fileReplacementDto, CancellationToken ct)
     {
-        Mediator.Publish(new HaltScanMessage("Download"));
+        Mediator.Publish(new HaltScanMessage(nameof(DownloadFiles)));
         try
         {
             await DownloadFilesInternal(gameObject, fileReplacementDto, ct).ConfigureAwait(false);
@@ -56,22 +77,22 @@ public partial class FileDownloadManager : DisposableMediatorSubscriberBase
         finally
         {
             Mediator.Publish(new DownloadFinishedMessage(gameObject));
-            Mediator.Publish(new ResumeScanMessage("Download"));
+            Mediator.Publish(new ResumeScanMessage(nameof(DownloadFiles)));
         }
     }
 
     protected override void Dispose(bool disposing)
     {
         CancelDownload();
-        base.Dispose(disposing);
-    }
-
-    public static void MungeBuffer(Span<byte> buffer)
-    {
-        for (int i = 0; i < buffer.Length; ++i)
+        foreach (var stream in _activeDownloadStreams)
         {
-            buffer[i] ^= 42;
+            try
+            {
+                stream.Dispose();
+            }
+            catch { }
         }
+        base.Dispose(disposing);
     }
 
     private static byte MungeByte(int byteOrEof)
@@ -86,8 +107,8 @@ public partial class FileDownloadManager : DisposableMediatorSubscriberBase
 
     private static (string fileHash, long fileLengthBytes) ReadBlockFileHeader(FileStream fileBlockStream)
     {
-        List<char> hashName = new();
-        List<char> fileLength = new();
+        List<char> hashName = [];
+        List<char> fileLength = [];
         var separator = (char)MungeByte(fileBlockStream.ReadByte());
         if (separator != '#') throw new InvalidDataException("Data is invalid, first char is not #");
 
@@ -121,7 +142,7 @@ public partial class FileDownloadManager : DisposableMediatorSubscriberBase
         Logger.LogDebug("Downloading {requestUrl} for request {id}", requestUrl, requestId);
         try
         {
-            response = await _orchestrator.SendRequestAsync(HttpMethod.Get, requestUrl, ct).ConfigureAwait(false);
+            response = await _orchestrator.SendRequestAsync(HttpMethod.Get, requestUrl, ct, HttpCompletionOption.ResponseHeadersRead).ConfigureAwait(false);
             response.EnsureSuccessStatusCode();
         }
         catch (HttpRequestException ex)
@@ -133,6 +154,7 @@ public partial class FileDownloadManager : DisposableMediatorSubscriberBase
             }
         }
 
+        ThrottledStream? stream = null;
         try
         {
             var fileStream = File.Create(tempPath);
@@ -142,7 +164,11 @@ public partial class FileDownloadManager : DisposableMediatorSubscriberBase
                 var buffer = new byte[bufferSize];
 
                 var bytesRead = 0;
-                while ((bytesRead = await (await response.Content.ReadAsStreamAsync(ct).ConfigureAwait(false)).ReadAsync(buffer, ct).ConfigureAwait(false)) > 0)
+                var limit = _orchestrator.DownloadLimitPerSlot();
+                Logger.LogTrace("Starting Download of {id} with a speed limit of {limit}", requestId, limit);
+                stream = new ThrottledStream(await response.Content.ReadAsStreamAsync(ct).ConfigureAwait(false), limit);
+                _activeDownloadStreams.Add(stream);
+                while ((bytesRead = await stream.ReadAsync(buffer, ct).ConfigureAwait(false)) > 0)
                 {
                     ct.ThrowIfCancellationRequested();
 
@@ -170,14 +196,24 @@ public partial class FileDownloadManager : DisposableMediatorSubscriberBase
             }
             throw;
         }
+        finally
+        {
+            if (stream != null)
+            {
+                _activeDownloadStreams.Remove(stream);
+                await stream.DisposeAsync().ConfigureAwait(false);
+            }
+        }
     }
 
     private async Task DownloadFilesInternal(GameObjectHandler gameObjectHandler, List<FileReplacementData> fileReplacement, CancellationToken ct)
     {
         Logger.LogDebug("Download start: {id}", gameObjectHandler.Name);
 
-        List<DownloadFileDto> downloadFileInfoFromService = new();
-        downloadFileInfoFromService.AddRange(await FilesGetSizes(fileReplacement.Select(f => f.Hash).Distinct(StringComparer.Ordinal).ToList(), ct).ConfigureAwait(false));
+        List<DownloadFileDto> downloadFileInfoFromService =
+        [
+            .. await FilesGetSizes(fileReplacement.Select(f => f.Hash).Distinct(StringComparer.Ordinal).ToList(), ct).ConfigureAwait(false),
+        ];
 
         Logger.LogDebug("Files with size 0 or less: {files}", string.Join(", ", downloadFileInfoFromService.Where(f => f.Size <= 0).Select(f => f.Hash)));
 
@@ -218,9 +254,10 @@ public partial class FileDownloadManager : DisposableMediatorSubscriberBase
             // let server predownload files
             var requestIdResponse = await _orchestrator.SendRequestAsync(HttpMethod.Post, MareFiles.RequestEnqueueFullPath(fileGroup.First().DownloadUri),
                 fileGroup.Select(c => c.Hash), token).ConfigureAwait(false);
-            Logger.LogDebug("Sent request for {n} files on server {uri} with result {result}", fileGroup.Count(), fileGroup.First().DownloadUri, requestIdResponse.Content.ReadAsStringAsync().Result);
+            Logger.LogDebug("Sent request for {n} files on server {uri} with result {result}", fileGroup.Count(), fileGroup.First().DownloadUri,
+                await requestIdResponse.Content.ReadAsStringAsync(token).ConfigureAwait(false));
 
-            Guid requestId = Guid.Parse(requestIdResponse.Content.ReadAsStringAsync().Result.Trim('"'));
+            Guid requestId = Guid.Parse((await requestIdResponse.Content.ReadAsStringAsync().ConfigureAwait(false)).Trim('"'));
 
             Logger.LogDebug("GUID {requestId} for {n} files on server {uri}", requestId, fileGroup.Count(), fileGroup.First().DownloadUri);
 
@@ -234,15 +271,15 @@ public partial class FileDownloadManager : DisposableMediatorSubscriberBase
                 {
                     try
                     {
-                        if (!_downloadStatus.ContainsKey(fileGroup.Key)) return;
-                        _downloadStatus[fileGroup.Key].TransferredBytes += bytesDownloaded;
+                        if (!_downloadStatus.TryGetValue(fileGroup.Key, out FileDownloadStatus? value)) return;
+                        value.TransferredBytes += bytesDownloaded;
                     }
                     catch (Exception ex)
                     {
                         Logger.LogWarning(ex, "Could not set download progress");
                     }
                 });
-                await DownloadAndMungeFileHttpClient(fileGroup.Key, requestId, fileGroup.ToList(), blockFile, progress, token).ConfigureAwait(false);
+                await DownloadAndMungeFileHttpClient(fileGroup.Key, requestId, [.. fileGroup], blockFile, progress, token).ConfigureAwait(false);
             }
             catch (OperationCanceledException)
             {
@@ -274,7 +311,7 @@ public partial class FileDownloadManager : DisposableMediatorSubscriberBase
                     try
                     {
                         Logger.LogDebug("Found file {file} with length {le}, decompressing download", fileHash, fileLengthBytes);
-                        var fileExtension = fileReplacement.First(f => string.Equals(f.Hash, fileHash, StringComparison.OrdinalIgnoreCase)).GamePaths[0].Split(".").Last();
+                        var fileExtension = fileReplacement.First(f => string.Equals(f.Hash, fileHash, StringComparison.OrdinalIgnoreCase)).GamePaths[0].Split(".")[^1];
 
                         byte[] compressedFileContent = new byte[fileLengthBytes];
                         _ = await fileBlockStream.ReadAsync(compressedFileContent, token).ConfigureAwait(false);
@@ -299,7 +336,8 @@ public partial class FileDownloadManager : DisposableMediatorSubscriberBase
             finally
             {
                 _orchestrator.ReleaseDownloadSlot();
-                fileBlockStream?.Dispose();
+                if (fileBlockStream != null)
+                    await fileBlockStream.DisposeAsync().ConfigureAwait(false);
                 File.Delete(blockFile);
             }
         }).ConfigureAwait(false);
@@ -313,7 +351,7 @@ public partial class FileDownloadManager : DisposableMediatorSubscriberBase
     {
         if (!_orchestrator.IsInitialized) throw new InvalidOperationException("FileTransferManager is not initialized");
         var response = await _orchestrator.SendRequestAsync(HttpMethod.Get, MareFiles.ServerFilesGetSizesFullPath(_orchestrator.FilesCdnUri!), hashes, ct).ConfigureAwait(false);
-        return await response.Content.ReadFromJsonAsync<List<DownloadFileDto>>(cancellationToken: ct).ConfigureAwait(false) ?? new List<DownloadFileDto>();
+        return await response.Content.ReadFromJsonAsync<List<DownloadFileDto>>(cancellationToken: ct).ConfigureAwait(false) ?? [];
     }
 
     private void PersistFileToStorage(string fileHash, string filePath)
@@ -321,7 +359,7 @@ public partial class FileDownloadManager : DisposableMediatorSubscriberBase
         var fi = new FileInfo(filePath);
         Func<DateTime> RandomDayInThePast()
         {
-            DateTime start = new(1995, 1, 1);
+            DateTime start = new(1995, 1, 1, 1, 1, 1, DateTimeKind.Local);
             Random gen = new();
             int range = (DateTime.Today - start).Days;
             return () => start.AddDays(gen.Next(range));
@@ -333,11 +371,11 @@ public partial class FileDownloadManager : DisposableMediatorSubscriberBase
         try
         {
             var entry = _fileDbManager.CreateCacheEntry(filePath);
-            if (!string.Equals(entry?.Hash, fileHash, StringComparison.OrdinalIgnoreCase))
+            if (entry != null && !string.Equals(entry.Hash, fileHash, StringComparison.OrdinalIgnoreCase))
             {
-                Logger.LogError("Hash mismatch after extracting, got {hash}, expected {expectedHash}, deleting file", entry?.Hash, fileHash);
+                Logger.LogError("Hash mismatch after extracting, got {hash}, expected {expectedHash}, deleting file", entry.Hash, fileHash);
                 File.Delete(filePath);
-                _fileDbManager.RemoveHashedFile(entry);
+                _fileDbManager.RemoveHashedFile(entry.Hash, entry.PrefixedFilePath);
             }
         }
         catch (Exception ex)
